Fixed bullet life span bugs, tileset grid fixed indefinitely.

+++ TileBasedPlayer20172018/Sprites/AnimatedSprite.cs
@@ -170,7 +170,7 @@ namespace AnimatedSprite
         public bool TileCollision(Tile other)
         {
             Rectangle myBound = new Rectangle((int)this.PixelPosition.X, (int)this.PixelPosition.Y, this.FrameWidth, this.FrameHeight);
-            Rectangle otherBound = new Rectangle((int)other.X * other.TileWidth, (int)other.Y * other.TileHeight, other.TileWidth, other.TileHeight);
+            Rectangle otherBound = new Rectangle((int)other.X * (int)other.TileWidth, (int)other.Y * (int)other.TileHeight, (int)other.TileWidth, (int)other.TileHeight);
 
             return myBound.Intersects(otherBound);
         }
@@ -178,7 +178,7 @@ namespace AnimatedSprite
         public bool InTile(Tile other)
         {
             Rectangle myBound = new Rectangle((int)this.PixelPosition.X, (int)this.PixelPosition.Y, this.FrameWidth, this.FrameHeight);
-            Rectangle otherBound = new Rectangle((int)other.X * other.TileWidth, (int)other.Y * other.TileHeight, other.TileWidth, other.TileHeight);
+            Rectangle otherBound = new Rectangle((int)other.X * (int)other.TileWidth, (int)other.Y * (int)other.TileHeight, (int)other.TileWidth, (int)other.TileHeight);
 
             return otherBound.Contains(myBound);
         }
